Rework TOTP structure
totps and totp_setups are now stored in the same table.

totps are now keyed by project_id+user_id+type  where the type is an arbitrary
string (and defaults to empty). The purpose of a type is to allow 1 user to
have multiple TOTP (say, one for login, and one for some other features).

Added TOTP verify API